Extract macro editor and refactor bindings
Move macro editing UI and related setup into a new macroeditor.lua module and simplify binding GUI logic to use it.
